Fix a bug in Category and Host classes
When the reports attribute of either class was
null and the method addReport was called it would
try to call the contains member method on a null value.

The addReport method now initialize the reports
attribute to an empty ArrayCollection before
the contains method is called if reports is null.